refactor(frontend): move undo-delete registry into UndoDeleteProvider and useUndoDelete hook
